requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
dacite==1.8.1
dataclasses-json==0.6.4
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
executing==2.0.1
faiss-cpu==1.8.0
fiona==1.9.5
fonttools==4.49.0
frozendict==2.4.0
frozenlist==1.4.1
geopandas==0.14.3
gitdb==4.0.11
GitPython==3.1.42
greenlet==3.0.3
h11==0.14.0
html5lib==1.1
htmlmin==0.1.12
httpcore==1.0.4
httpx==0.27.0
idna==3.6
ImageHash==4.3.1
ipykernel==6.29.3
ipython==8.22.1
ipywidgets==8.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_widgets==3.0.10
kaggle==1.6.11
kiwisolver==1.4.5
langchain==0.1.10
langchain-community==0.0.31
langchain-core==0.1.38
langchain-text-splitters==0.0.1
langsmith==0.1.13
llvmlite==0.42.0
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.0
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
MouseInfo==0.1.3
multidict==6.0.5
multimethod==1.11.2
multitasking==0.0.11
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.59.1
numpy==1.26.4
openai==1.13.3
orjson==3.9.15
packaging==23.2
pandas==2.2.1
parso==0.8.3
patsy==0.5.6
peewee==3.17.1
phik==0.12.4
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyarrow==15.0.1
pydantic==2.6.3
pydantic_core==2.16.3
pydeck==0.8.1b0
PyGetWindow==0.0.9
Pygments==2.17.2
PyMsgBox==1.0.9
pyparsing==3.1.1
pypdf==4.1.0
pyperclip==1.8.2
pyproj==3.6.1
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.8.2
python-dotenv==1.0.1
python-slugify==8.0.4
pytweening==1.2.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
scikit-learn==1.4.1.post1
scipy==1.11.4
seaborn==0.12.2
setuptools==69.1.1
shapely==2.0.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.27
stack-data==0.6.3
statsmodels==0.14.1
streamlit==1.32.0
tenacity==8.2.3
text-unidecode==1.3
threadpoolctl==3.4.0
tiktoken==0.6.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
typeguard==4.2.1
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.2.1
visions==0.7.6
watchdog==4.0.0
wcwidth==0.2.13
webencodings==0.5.1
widgetsnbextension==4.0.10
wordcloud==1.9.3
yarl==1.9.4
